requirements.txt
numpy
streamlit
deepface
fastapi
uvicorn
opencv-python
python-multipart
openface
pytesseract
Pillow